Implement generating overview page.
This change adds a new feature to generate an overview page for all of the rule
sets in the generated documentation. This is particularly useful for HTML
documentation and allows users to directly drop the generated HTML docs onto
a webserver and serve the docs as a standalone website.

This change also adds a link_ext attribute, which allows for customizing the
file extension used for links in the generated documentation since some
webservers serve Markdown pages with the .md file extension rather than .html.

Fixes #24