Implementing extracting symbols from load()s and adding them to stubs.
This change adds LoadExtractor, which walks the AST of the .bzl file and
extracts information from the load() statements about symbols that are loaded
from other .bzl files. These symbols are then added to the set of global stubs
in RuleDocExtractor so that, as a first step, if any of these symbols loaded
from other .bzl files are referenced in the global scope, Skydoc would not
raise a NameError.